zayballl

+++ database/migrations/2024_04_15_142437_add_tenant_id_to_tariff_payment_table.php
@@ -8,14 +8,16 @@ return new class extends Migration {
 
     public function up(): void
     {
-        Schema::connection('mysql')->table('tariff_payment', function (Blueprint $table) {
-            $table->string('tenant_id');
-            $table->foreign('tenant_id')
-                ->references('id')
-                ->on('tenants')
-                ->onUpdate('cascade')
-                ->onDelete('cascade');
-        });
+        if (!column_exists('tariff_payment', 'tenant_id', 'mysql')) {
+            Schema::connection('mysql')->table('tariff_payment', function (Blueprint $table) {
+                $table->string('tenant_id');
+                $table->foreign('tenant_id')
+                    ->references('id')
+                    ->on('tenants')
+                    ->onUpdate('cascade')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     public function down(): void
